% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nefsc_groundfish_access.R
\name{add_area_stratification}
\alias{add_area_stratification}
\title{Add Survey Area Stratified Abundances and Biomasses}
\usage{
add_area_stratification(
  survdat_weights,
  include_epu = F,
  box_location = "root|cloudstorage"
)
}
\arguments{
\item{survdat_weights}{Input dataframe, produced by add_lw_info}

\item{include_epu}{Flag for calculating the EPU rates in addition to the stratum regions we use.}

\item{box_location}{String indicating value to pass to `boxpath_switch`}
}
\value{
survdat df with fields for area stratified values
}
\description{
Take the survdat data which has been cleaned with gmRi::gmri_survdat_prep, that is
paired with length weight relationships using gmRi::add_lw_info and estimate the area stratified
catch rates and their expected abundances and biomasses when those catch rates are applied to
the total areas of their respective strata.

Area-stratified catch rates are calculated independently for each species, every year, within
each strata.

Constants for the area-towed and the catchability coefficient are as followed:
- Area covered by an albatross standard tow in km2 = 0.0384
- Catchability coefficient - ideally should change for species guilds: q = 1
}
\examples{
# not run
# add_area_stratification(survdat_weights = survdat_lw, include_epu = F)
}
