% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nefsc_groundfish_access.R
\name{add_epu_info}
\alias{add_epu_info}
\title{Add EPU Details to Station Info}
\usage{
add_epu_info(trawldat, box_location = "root|cloudstorage")
}
\arguments{
\item{trawldat}{Survdat data with decdeg_beglat and decdeg_beglon coordinates}

\item{box_location}{String indicating value to pass to `boxpath_switch`}
}
\value{
survdat dataframe containing epu overlap information
}
\description{
Uses Sean Lucey's Poststrat function to add the correct EPU designation
to the survey station locations.

Uses Sean Lucey's "poststrat" function to overlay EPU's. Tried to
replace it with sf steps, but it wasn't worth the effort. This function exists to
reduce clutter in the area stratification function
}
\examples{
#' # not run
# add_epu_info(trawldat = survdat_lw)
}
