% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{as_fahrenheit}
\alias{as_fahrenheit}
\title{Convert to Fahrenheit from Celsius}
\usage{
as_fahrenheit(temp_c, data_type = "temperature")
}
\arguments{
\item{temp_c}{Values of temperature or anomalies in Deg C}

\item{data_type}{T/F Toggle to convert temperatures or anomalies}
}
\value{
temperature in fahrenheit
}
\description{
Convert temperature in Celsius to Fahrenheit.
Toggle for anomaly data (does not add 32).
}
\examples{
as_fahrenheit(c(15,20), data_type = "temperature")
}
