% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_paths_allyn.R
\name{box_path}
\alias{box_path}
\title{Box Path Generator}
\usage{
box_path(box_group = NULL, subfolder = NULL)
}
\arguments{
\item{box_group}{The top level directory name within Box}

\item{subfolder}{Any subsequent sub-directory location you wish to access}
}
\value{
returns a path to box
}
\description{
Create OS agnostic paths to different locations on box.
}
\examples{
oisst_path <- box_path(box_group = "res", subfolder = "OISST/oisst_mainstays")
shapefile_path <- box_path(box_group = "res", subfolder = "Shapefiles")
}
