% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_paths_allyn.R
\name{boxpath_switch}
\alias{boxpath_switch}
\title{Box Path Switcher for Mac Storage Location}
\usage{
boxpath_switch(box_location = "base")
}
\arguments{
\item{box_location}{String where Mac OS system mounted path to box. Accepts "cloudstorage" &
"mojave" to be backwards compatible with code that mistook this for a mojave issue,
all other values will use `box_path` and not `cs_path`}
}
\value{
Returns path function based on operating system
}
\description{
Toggles the root location of Box cloud storage based on whether a box was mounted
under libraries or under cloudstorage
}
\examples{
boxpath_fun <- boxpath_switch(box_location = "cloudstorage")
}
