% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oisst_data_funs.R
\name{calc_daily_climatologies}
\alias{calc_daily_climatologies}
\title{Vectorized Daily OISST 30-Year Climatology Means}
\usage{
calc_daily_climatologies(stack_in, projection_crs = 26919, anom_period = NULL)
}
\arguments{
\item{stack_in}{Raster stack of daily SST Measurements from OISST dataset}

\item{projection_crs}{The coordinate reference system number/proj4string for the desired output}

\item{anom_period}{Character vector indicating start and end years to calculate climatology for. NULL defaults to 1982-2011.}
}
\value{

}
\description{
Generate Daily 30-year climatologies for a given area.

Returns a raster stack with mean SST for every day in the year over the desired time window.
}
