% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allyn_sharedboxpaths.R
\name{enter_user_name_func}
\alias{enter_user_name_func}
\title{Check & Enter User Name}
\usage{
enter_user_name_func()
}
\value{
user.name User name
}
\description{
Prompts user name entry for determining box location on local machine.
(needs final check for "windows" machine users). Mac users do not need this step.
}
\examples{
#Not Run
#enter_user_name()

}
