% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oisst_data_funs.R
\name{env_data_extract}
\alias{env_data_extract}
\title{Satellite Data Extraction Function}
\usage{
env_data_extract(
  data.set = "OISST",
  dates = NULL,
  box = c(-77, -60, 35, 46),
  out.dir = here::here("data", "OISST_thredd_test"),
  mask = NULL
)
}
\arguments{
\item{data.set}{Env data to extract (options = ERSST, OISST, MURSST)}

\item{dates}{If !NULL, subset full time series to specific dates. Dates should be specified as dates = c("YYYY-MM-DD", "YYYY-MM-DD")}

\item{box}{If !NULL, crop rasters to sepcific box faster downloading and processing. Box should be specified as box = c(xmin, xmax, ymin, ymax).}

\item{out.dir}{Directory to store resulting raster stack. Note, this will overwrite any rasters with the existing name.}

\item{mask}{optional mask to trim data with}
}
\value{
stack.out Stack of daily OISST files and also saves the raster stack as a .grd file in the output directory specified by out.dir
}
\description{
This function accesses webhosted satellite data and then downloads a subset of the data based
on dates and long/lat bounding box. After downloading the data, the function processes it and saves it as
one raster stack file. Installs "ncdf4" and "raster" packages if not installed.
}
