% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oisst_data_funs.R
\name{fix_raster}
\alias{fix_raster}
\title{Raster from Array}
\usage{
fix_raster(x, lons.use, lats.use, x.min.use, x.max.use, y.min.use, y.max.use)
}
\arguments{
\item{x}{matrix. As currently implemented, x is an individiaul time slice from an x, y, z array.}

\item{lons.use}{Longitudes, extracted using ncvar_get}

\item{lats.use}{Latitudes, extracted using ncvar_get}

\item{x.min.use}{Bounding box x minimum}

\item{x.max.use}{Bounding box x maximum}

\item{y.min.use}{Bounding box y minimum}

\item{y.max.use}{Bounding box y maximum}
}
\value{
rast.out Correctly trimmed and oriented raster object
}
\description{
Helper function to convert an array (x, y, z)
into raster layers with the correct dimensions and orientation for spatial analyses.
Typically used when working from netcdf to raster.
}
