% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{floor_decade}
\alias{floor_decade}
\title{Floor Decade}
\usage{
floor_decade(year_vector, return_class = "factor")
}
\arguments{
\item{year_vector}{Vector of years.}

\item{return_class}{Optional flag to return vector formatted as a factor, alternative is numeric.}
}
\value{
returns vector of values representing the decade each year occurred in.
}
\description{
Returns the decade for vector of years.
}
\examples{
# not run
}
