% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes.R
\name{geom_logo}
\alias{geom_logo}
\title{Ggplot2 Raster Logo as Geom Layer}
\usage{
geom_logo(
  logo = NULL,
  x_npc = 0.9,
  y_npc = 0.925,
  logo_height = 0.06,
  height_units = "npc",
  ...
)
}
\arguments{
\item{logo}{Any R object that can be coerced to a raster object. Ex. class of "magick-image"}

\item{x_npc}{x positioning for logo in normalized parent coordinates 0-1 left to right}

\item{y_npc}{y positioning for logo in normalized parent coordinates 0-1 bottom to top}

\item{logo_height}{Desired height of the logo}

\item{height_units}{Units to set the height (and width) for the logo.
Default is NPC i.e. a fraction of the plot panel height}

\item{...}{Extra control options passed to ggplot2::annotation_custom()}
}
\value{
Returns a  ggplot geom layer
}
\description{
Ggplot2 Raster Logo as Geom Layer
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(cyl, hp)) + geom_point() + geom_logo(logo_height = 0.06)
}
