% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oisst_mainstays_data_access.R
\name{get_region_names}
\alias{get_region_names}
\title{Lookup Tool for Timeseries & Mask Locations}
\usage{
get_region_names(region_group = "gmri_sst_focal_areas")
}
\arguments{
\item{region_group}{Name of the group of shapefiles of interest:
"gmri_sst_focal_areas", "lme", "nmfs_trawl_regions", ""}
}
\value{
Vector of names included in the selectedd region group
}
\description{
This is a hard-coded catalog of what resources have been
processed as part of the oisst-mainstays processing steps.

Names are presented in the way that they are saved. This function then
is used as a unified lookup source for accessing timeseries or shapefiles.
}
\examples{
lme_areas <- get_region_names("lme")
}
