% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oisst_mainstays_data_access.R
\name{get_timeseries_paths}
\alias{get_timeseries_paths}
\title{Timeseries and Shapefile Path Look-Up}
\usage{
get_timeseries_paths(region_group, os.use = "unix", user.name = NULL)
}
\arguments{
\item{region_group}{Name of the region group used to fetch region names from
gmRi::get_region_names(): Options: gmri_sst_focal_areas, nelme_regions, nmfs_trawl_regions, lme}

\item{os.use}{Operating system setting passed to gmRi::shared.path()}

\item{user.name}{Optional configuration for Windows users trying to use shared.path}
}
\value{
Returns list of file paths to both the shapefiles and their sst timeseries
}
\description{
Returns the full file paths for a collection of regionally masked timeseries
for sea surface temperature, or to the shapefiles used as the masks.

This is a parallel to the python functions used to create them and ensure consistent names.
}
\examples{
# Not Run:
# r_group <- "lme"
# get_timeseries_paths(region_group = r_group, region_list = get_region_names(r_group))
}
