% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmRi-package.R
\docType{package}
\name{gmRi-package}
\alias{gmRi}
\alias{gmRi-package}
\title{gmRi: Utility Functions for GMRI Research Team}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Common routines and utilities for performing various 
  research tasks and implementing GMRI branding. Currently this package 
  adds ggplot2 palette tools to access official Gulf of Maine Research 
  Institute colors.
}
\author{
\strong{Maintainer}: Adam Kemberling \email{akemberling@gmri.org} (\href{https://orcid.org/0000-0003-0402-0829}{ORCID})

}
\keyword{internal}
