% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmri_palettes.R
\name{gmri_cols}
\alias{gmri_cols}
\title{Retrieve GMRI Hex Color(s)}
\usage{
gmri_cols(...)
}
\arguments{
\item{...}{Character names of official GMRI colors}
}
\description{
Return hex codes in return for named GMRI colors.
Available options are: seafood purple, midnight blue, dark blue, gmri blue,
ecosystems cyan, blue economy teal, moss green, warm yellow, lv orange, climate
change red, light gray, & dark gray.
}
\examples{

#Pull a single Hex code
gmri_cols("gmri blue")

#Multiple colors
gmri_cols("gmri blue", "moss green")

#Get the whole list of colors
gmri_cols()

#Add select colors to ggplot3 functions
ggplot2::ggplot(mtcars, ggplot2::aes(hp, mpg)) +
  ggplot2::geom_point(color = gmri_cols("gmri blue"), size = 4, alpha = .8)

}
