% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{\%not in\%}
\alias{\%not in\%}
\title{Not In}
\usage{
x \%not in\% table
}
\arguments{
\item{x}{vector that you wish to check for matches}

\item{table}{Table or vector containing strings that you wish to check x against.}
}
\value{

}
\description{
Negation of the `%in%` function.
}
