% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_stylesheets.R
\name{insert_gmri_footer}
\alias{insert_gmri_footer}
\title{Insert HTML Header File from gmRi package files - Rmd Implementation}
\usage{
insert_gmri_footer(footer_file = "akemberling_gmri_footer.html")
}
\arguments{
\item{footer_file}{String indicating desired css filename.}
}
\value{
HTML text to insert as is
}
\description{
Insert HTML Header File from gmRi package files - Rmd Implementation
}
\examples{
# Not Run use_gmri_footer(gmri_rmarkdown.css)
}
