% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_stylesheets.R
\name{insert_gmri_header}
\alias{insert_gmri_header}
\title{Insert HTML Header File from gmRi package files - Rmd Implementation}
\usage{
insert_gmri_header(header_file = "gmri_logo_header.html")
}
\arguments{
\item{header_file}{String indicating desired css filename.}
}
\value{
HTML text inserted as-is
}
\description{
Insert HTML Header File from gmRi package files - Rmd Implementation
}
\examples{
# Not Run use_gmri_header(gmri_rmarkdown.css)
}
