% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oisst_mainstays_data_access.R
\name{load_global_oisst}
\alias{load_global_oisst}
\title{Access OISST Mainstays Arrays}
\usage{
load_global_oisst(
  resource = c("raw", "climatology82", "climatology91", "anomalies", "warming rates"),
  year_range = seq(2010, 2020, 1),
  box_location = "default_boxpath|cloudstorage"
)
}
\arguments{
\item{resource}{Name of the global extent resource, choices are raw, climatology, warming rates,
and anomalies.}

\item{year_range}{optional vector of years for raw or anomalies data resources.}

\item{box_location}{String indicating value to pass to `boxpath_switch`}
}
\value{
resource_out Raster stack of the desired netcdf array
}
\description{
Access OISST Mainstays Arrays
}
\examples{
# Not run
# load_global_oisst(resource = "warming rates", year_range = NULL)
}
