% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oisst_data_funs.R
\name{make360}
\alias{make360}
\title{Longitude 180 to 360}
\usage{
make360(lon)
}
\arguments{
\item{lon}{Longitude in -180:180 degrees}
}
\value{
lon Longitude from 0-360 degrees
}
\description{
This is a simple function to translate negative longitudes
(measured on -180:180 scale) into 0-360, which is coordinate system used by some environmental datasets.
}
