% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{make_monthly}
\alias{make_monthly}
\title{Convert Daily Stack to Monthly Means}
\usage{
make_monthly(daily_ras)
}
\arguments{
\item{daily_ras}{Raster stack with daily time step}
}
\value{
Raster stack with monthly time step
}
\description{
Takes a stack of daily data with names of "XYYYY.MM.DD" and
returns monthly averages.
}
\examples{
# not run
}
