% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oisst_data_funs.R
\name{make_stack_dates}
\alias{make_stack_dates}
\title{Make Valid Raster Stack Dates}
\usage{
make_stack_dates(point_location_df, year_col, month_col, day_col)
}
\arguments{
\item{point_location_df}{Dataframe of point locations we wish to use to extract data from raster stack}

\item{year_col}{name of the column that contains year information (unquoted text)}

\item{month_col}{name of the column that contains the month information (unquoted text)}

\item{day_col}{name of the column that contains the day information (unquoted text)}
}
\value{
df_out Original dataframe with added "valid_dates" column
}
\description{
Make Valid Raster Stack Dates
}
\details{
Takes a dataframe object containing columns for the year, month, and day
and creates a new "valid_dates" column that will match the naming convention for
a raster stack object, i.e. names(raster_stack) which cannot start with a number
or contain spaces or underscores. Month and day values are left-padded with zeros
to maintain consistent string length.
}
