% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nefsc_groundfish_access.R
\name{make_survdat_occu}
\alias{make_survdat_occu}
\title{make SURVDAT tidy occupancy dataset}
\usage{
make_survdat_occu(survdat_clean, species_keep)
}
\arguments{
\item{survdat_clean}{Survdat data, after usual preparations are completed.
These include removal of old strata, labeling of areas of interest, and inclusion
of the annual effort in each.}

\item{species_keep}{Vector of character species names or vector of numeric species codes to filter SURVDAT['comname'] (if character), or SURVDAT['svspp'] (if numeric)}
}
\value{
tidy occupancy dataframe containing one row for every unique tow-species
}
\description{
make a tidy occupancy dataset from the cleaned SURVDAT dataset, where each row represents a unique tow - species - occupancy record for every species in species_keep. In creating this occupancy dataset, we take the total biomass or abundance across sex/sizes and impute "absence" observations, such that there is a record for every species in `species_keep` at every unique tow.
}
\examples{
# not run
# make_survdat_occu(survdat_clean = survdat_clean, species_keep = c("atlantic cod", "offshore hake"))
}
