% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes.R
\name{map_theme}
\alias{map_theme}
\title{GGplot2 Map Theme}
\usage{
map_theme(...)
}
\arguments{
\item{...}{Ellipse are passed to ggplot2::theme() function allowing users}
}
\value{
Returns ggtheme
}
\description{
A simple theme that works well for maps made with {sf}.
Black border all around, and redundant lat/lon labels removed.
}
\examples{
ggplot2::ggplot(mtcars) + map_theme()
}
