% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsf_okn_data_access.R
\name{nsf_access_paths}
\alias{nsf_access_paths}
\title{Connect to NSF OKN Demo Data Paths}
\usage{
nsf_access_paths(os.use = "unix", user.name = "not applicable I use a mac")
}
\arguments{
\item{os.use}{String flag indicating what operating system the user is currently using. Options are "unix" for mac users or "windows".}

\item{user.name}{User name for navigating root directory structure on windows.}
}
\value{
res_paths List containing user-specific paths
}
\description{
Connect to NSF OKN Demo Data Paths
}
\examples{
# Not Run
}
