% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oisst_mainstays_data_access.R
\name{oisst_access_timeseries}
\alias{oisst_access_timeseries}
\title{Access Regional Timeseries from OISSTv2 Mainstays}
\usage{
oisst_access_timeseries(
  region_family = "gmri",
  poly_name = "gulf of maine",
  box_location = "default_boxpath|cloudstorage"
)
}
\arguments{
\item{region_family}{Identify the family of shapefiles that you are interested in.
Choices = "LME", "nmfs trawl regions", "gmri focus areas"}

\item{poly_name}{String Identifying the shapefile name that was used as mask.
Used to build file name.}

\item{box_location}{String indicating value to pass to `boxpath_switch`}
}
\value{
Time series dataframe for the selected region.
}
\description{
Tool for accessing oisst regional timeseries from Box.
Regionally masked timeseries have been pre-processed and stored on box. Quick access
is available here.

This function is used primarily as a pass-through function for oisst_access_timeseries(),
which relies on it to look-up file paths.

Simply specify your personal path to the oisst_mainstays folder,
the group of polygons your area is in, and the name of the region itself.
}
\examples{
#Not run:

# box_paths <- gmRi::research_access_paths(os.use = "unix",
#                                          user.name = "not applicable")
# agulhas <- oisst_access_timeseries(
#    region_family = "lme",
#    poly_name = "agulhas current",
#    box_location = "default_boxpath|cloudstorage)
}
