% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oisst_data_funs.R
\name{oisst_period_means}
\alias{oisst_period_means}
\title{OISST Time-Period Means}
\usage{
oisst_period_means(stack_in, projection_crs = 26919, time_res_df)
}
\arguments{
\item{stack_in}{Raster stack of daily SST Measurements from OISST dataset}

\item{projection_crs}{The coordinate reference system number/proj4string for the desired output}

\item{time_res_df}{Dataframe detailing the time-period structure you wish to calculate SST means for}
}
\value{

}
\description{
Generate average annual SST measurements from the daily OISST
raster brick using custom date windows. Takes the daily OISST mean temperature
raster stack and a dataframe outlining period start and end dates as inputs.

Returns a raster stack with mean SST For every year for each intra-annual time-period
specified.

Useful for getting means across a specific time window within a year, like a month or season.
}
