% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oisst_mainstays_data_access.R
\name{oisst_window_load}
\alias{oisst_window_load}
\title{Access OISST via Data Window of lat/lon/time}
\usage{
oisst_window_load(
  data_window,
  anomalies = FALSE,
  box_location = "default_boxpath|cloudstorage"
)
}
\arguments{
\item{data_window}{dataframe with columns for lat, lon, & time indicating the extent of
the data desired.}

\item{anomalies}{Boolean indication of whether to return observed sst or anomalies.
Default = TRUE.}

\item{box_location}{String indicating value to pass to `boxpath_switch`}
}
\value{
Raster stack of OISSTv2 data using desired dimensions to crop
}
\description{
Returns a list of OISST raster stacks that
has been clipped to the desired lat/lon/time extent. Each list element
is named according to year, and contains all daily measurements within the
desired window of time, and clipped to the desired lat/lon extent.


Useful for loading OISST observations from box, but only for a particular area and/or time.
}
\examples{
#not run
# oisst_path <- box_path(box_group = "res", subfolder = "OISST/oisst_mainstays")
# data_window <- data.frame(lon = c(-72, -65),
#                           lat = c(42,44),
#                           time = as.Date(c("2016-08-01", "2020-12-31")))

}
