% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oisst_mainstays_data_access.R
\name{okn_access_timeseries}
\alias{okn_access_timeseries}
\title{Access Regional Timeseries from OKN Demo Data}
\usage{
okn_access_timeseries(
  oisst_path,
  sat_source = "oisst",
  region_family = c("nmfs trawl regions", "lme", "gmri focus areas"),
  poly_name = "gulf of maine"
)
}
\arguments{
\item{oisst_path}{Personal path to ~Box/RES_Data/OISST/oisst_mainstays}

\item{sat_source}{String indicating satellite data source.}

\item{region_family}{Identify the family of shapefiles that you are interested in. Choices = "LME", "Groundfish Regions"}
}
\value{

}
\description{
Tool for accessing NSF OKN Demo Data Products from Box. Global time series,
as well as regionally masked timeseries have been pre-processed and stored on box. Wuick access
is available here.

Simply name the satellite data source, the desired extent or mask, and the function will return the corresponding table.
}
\examples{
box_paths <- gmRi::research_access_paths(os.use = "unix", user.name = "not applicable I use a mac")
agulhas_timeseries <- okn_access_timeseries(okn_path = box_paths$oisst_mainstays, sat_source = "oisst", region_family = "lme", poly_name = "agulhas current")
}
