% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_paths_allyn.R
\name{os_fun_switch}
\alias{os_fun_switch}
\title{Mac OS Box Path Switch}
\usage{
os_fun_switch(mac_os = "pre_mojave")
}
\arguments{
\item{mac_os}{String indicating Mac OS system. Only valid value is "mojave", all other values
will use `box_path` and not `cs_path`}
}
\value{

}
\description{
Toggles the root location of Box cloud storage based on whether a user is running
}
