% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_palettes.R
\name{scale_color_gmri}
\alias{scale_color_gmri}
\title{GMRI-ggplot2 Scale Color Palettes}
\usage{
scale_color_gmri(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in gmri_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
GMRI-ggplot2 Scale Color Palettes
}
\examples{
#Color by discrete variable using default palette
ggplot2::ggplot(iris, ggplot2::aes(Sepal.Width, Sepal.Length, color = Species)) +
 ggplot2::geom_point(size = 4) +
 scale_color_gmri()

#Color by numeric variable with cool palette
ggplot2::ggplot(iris, ggplot2::aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
  ggplot2::geom_point(size = 4, alpha = .6) +
  scale_color_gmri(discrete = FALSE, palette = "cool")

}
