% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmri_palettes.R
\name{scale_fill_gmri}
\alias{scale_fill_gmri}
\title{GMRI-ggplot2 Scale Fill Palettes}
\usage{
scale_fill_gmri(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in gmri_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
GMRI-ggplot2 Scale Fill Palettes
}
\examples{
#Fill by discrete variable with different palette + remove legend (guide)
#ggplot2::ggplot(mpg, ggplot2::aes(manufacturer, fill = manufacturer)) +
#  ggplot2::geom_bar() +
#  ggplot2::theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
#  scale_fill_gmri(palette = "mixed", guide = "none")

}
