% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_paths_allyn.R
\name{shared.path}
\alias{shared.path}
\title{Access Shared Path from Box Drive}
\usage{
shared.path(
  os.use = "unix",
  group = c("RES Data", "Mills Lab", "Climate Change Ecology Lab", "NSF OKN", "root"),
  folder = "Functions/"
)
}
\arguments{
\item{os.use}{Character string defining operating system. Either "unix"/"mac" or "windows"}

\item{group}{Character string of top-level directory within box to open. Ex. Box/RES_Data
Either RES (things shared across the entire research department) or
Mills Lab (things shared within the Mills Lab). Additional shortcuts are NSF OKN
for knowledge graph data resources and root for the minimal path to box.}

\item{folder}{Sub-folder within box you wish to access}
}
\value{
path.out Path to shared folder. We will then use these paths within shared code to load shared data or source shared functions.
}
\description{
This function creates paths to shared Data and Functions folders stored either inside the Mills
Lab folder or outside the Mills Lab folder in other top level locations on box.
}
\examples{
# res.data.path<- shared.path(os.use = os.use, group = "RES", folder = "Data/")

# Not Run
# gom.shapefile<- st_read(paste(res.data.path, "Shapefiles/GoM_sf.shp", sep = ""))
}
