% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes.R
\name{theme_gmri}
\alias{theme_gmri}
\title{GMRI ggplot2 theme for blog-style plots}
\usage{
theme_gmri(
  base_size = 10,
  base_family = "sans",
  title_family = "sans",
  facet_color = "blue economy teal",
  ...
)
}
\arguments{
\item{base_size}{Base text size (font numbers)}

\item{base_family}{Base Text font family}

\item{title_family}{Title font family}

\item{facet_color}{Color label passed to gmRi::gmri_cols() to pick a color for facet tabs}

\item{...}{Additional arguments to pass on the fly using ggplot2::theme()}
}
\value{
Returns ggplot theme
}
\description{
GMRI ggplot2 theme for blog-style plots
}
\examples{
ggplot2::ggplot(mtcars) + theme_gmri()
}
