% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes.R
\name{theme_gmri_simple}
\alias{theme_gmri_simple}
\title{Simple ggplot theme for cleaner look}
\usage{
theme_gmri_simple(
  base_size = 10,
  base_family = "Avenir",
  title_family = "Avenir",
  ...
)
}
\arguments{
\item{base_size}{Base text size (font numbers)}

\item{base_family}{Base Text font family}

\item{title_family}{Title font family}

\item{...}{Additional arguments to pass on the fly using ggplot2::theme()}
}
\value{
Returns ggplot theme
}
\description{
Simple ggplot theme for cleaner look
}
\examples{
ggplot2::ggplot(mtcars) + theme_gmri_simple()
}
