% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes.R
\name{theme_map_simple}
\alias{theme_map_simple}
\title{Simple Map theme with solid frame borders}
\usage{
theme_map_simple(fontfam = "Avenir", ...)
}
\arguments{
\item{fontfam}{font family to control the plotting of text}

\item{...}{}
}
\description{
Simple Map theme with solid frame borders
}
