% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_stylesheets.R
\name{use_gmri_footer}
\alias{use_gmri_footer}
\title{Insert HTML Header File from gmRi package files}
\usage{
use_gmri_footer(footer_file = "akemberling_gmri_footer.html")
}
\arguments{
\item{header_file}{String indicating desired css filename.}
}
\value{

}
\description{
Insert HTML Header File from gmRi package files
}
