% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_stylesheets.R
\name{use_gmri_style_rmd}
\alias{use_gmri_style_rmd}
\title{Use CSS File from gmRi package files - Rmd Implementation}
\usage{
use_gmri_style_rmd(css_file = "gmri_rmarkdown.css")
}
\arguments{
\item{css_file}{String indicating desired css filename.}
}
\value{

}
\description{
Use CSS File from gmRi package files - Rmd Implementation
}
\examples{
# Not Run use_gmri_style(gmri_rmarkdown.css)
}
