% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_stylesheets.R
\name{use_gmri_style_shiny}
\alias{use_gmri_style_shiny}
\title{Access GMRI Stylesheets - Shiny App Implementation}
\usage{
use_gmri_style_shiny(
  stylesheet = "gmri rmarkdown",
  header = "none",
  footer = "none"
)
}
\arguments{
\item{stylesheet}{String identifying which stylesheet to attach to Rmarkdown or Shiny App}

\item{header}{Optional string identifying html footer to attach, default is NULL, "gmri logo right" example given}

\item{footer}{Optional string identifying htmml footer to attach, default is NULL, "akemberling" example included}
}
\value{
Function prints the css and html elements attached
}
\description{
Access GMRI Stylesheets - Shiny App Implementation
}
\examples{
# Not Run
}
