% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nefsc_groundfish_access.R
\name{add_area_stratification}
\alias{add_area_stratification}
\title{Add Survey Area Stratified Abundances and Biomasses}
\usage{
add_area_stratification(
  survdat_weights,
  include_epu = F,
  mac_os = "pre_mojave"
)
}
\arguments{
\item{survdat_weights}{Input dataframe, produced by add_lw_info}

\item{include_epu}{Flag for calculating the EPU rates in addition to the stratum regions we use.}

\item{mac_os}{String indicating value to pass to `os_fun_switch`}
}
\value{
survdat df with fields for area stratified values
}
\description{
Take the survdat data paired with length weight relationships and
return estimates of area stratified catch rates and their expected abundances and
biomasses when applied to the total areas of stratum.
}
\examples{
# not run
# add_area_stratification(survdat_weights = survdat_lw, include_epu = F)
}
