% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nefsc_groundfish_access.R
\name{add_lw_info}
\alias{add_lw_info}
\title{Add species length weight information to cleaned SURVDAT trawl data}
\usage{
add_lw_info(survdat_clean, cutoff = FALSE, box_location = "root|cloudstorage")
}
\arguments{
\item{survdat_clean}{Survdat data, after usual preparations are completed.
These include removal of old strata, labeling of areas of interest, and inclusion
of the annual effort in each.}

\item{cutoff}{Flag for whether to remove species that have shown poor fit
when compared to biomass data from the nmfs survey.}

\item{box_location}{String indicating value to pass to `boxpath_switch`}
}
\value{
survdat dataframe containing paired length-weight relationship details for species
}
\description{
calculate expected biomass-at-length for species based on
published length-weight relationships.

Species are matched against a spreadsheet containing length and weight information from 2
sources. The first source is the length-weight relationships detailed in Wigley et al. 2003:
"Length-weight relationships for 74 fish species collected during NEFSC research vessel bottom
trawl surveys, 1992-99".

The second source for matching growth details to species is fishbase. These values are known to
be potentially less accurate or less regionally specific.

Pairings are first checked against the Wigley paper, and then by fishbase, to provide preference
to the more regionally focused source.
}
\examples{
# not run
# add_lw_info(survdat_clean = survdat_clean, cutoff = T)
}
