% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_paths_allyn.R
\name{cs_path}
\alias{cs_path}
\title{CloudStorage Box Path Generator}
\usage{
cs_path(box_group = NULL, subfolder = NULL)
}
\arguments{
\item{box_group}{The top level directory name within Box}

\item{subfolder}{Any subsequent sub-directory location you wish to access}
}
\value{
returns a path to box
}
\description{
Create Mac OS path to Box for Mojave users
}
\examples{
oisst_path <- cs_path(box_group = "res", subfolder = "OISST/oisst_mainstays")
shapefile_path <- cs_path(box_group = "res", subfolder = "Shapefiles")
}
