% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nefsc_groundfish_access.R
\name{gmri_survdat_prep}
\alias{gmri_survdat_prep}
\title{Load survdat file with standard data filters, keep all columns}
\usage{
gmri_survdat_prep(
  survdat = NULL,
  survdat_source = "most recent",
  box_location = "root|cloudstorage"
)
}
\arguments{
\item{survdat}{optional starting dataframe in the R environment to run through size spectra build.}

\item{survdat_source}{String indicating which survdat file to load from box}

\item{box_location}{String indicating value to pass to `boxpath_switch`}
}
\value{
Returns a dataframe filtered and tidy-ed for size spectrum analysis.
}
\description{
Processing function to prepare survdat data for size spectra analyses.
Options to select various survdat pulls, or provide your own as a dataframe
from the environment if available.
}
\examples{
# not run
# gmri_survdat_prep(survdat_source = "most recent")
}
