% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oisst_mainstays_data_access.R
\name{oisst_access_timeseries}
\alias{oisst_access_timeseries}
\title{Access Regional Timeseries from OISSTv2 Mainstays}
\usage{
oisst_access_timeseries(
  res_path,
  region_family = c("nmfs trawl regions", "lme", "gmri focus areas"),
  poly_name = "gulf of maine"
)
}
\arguments{
\item{res_path}{Personal path to ~Box/RES_Data/OISST/oisst_mainstays}

\item{region_family}{Identify the family of shapefiles that you are interested in.
Choices = "LME", "Groundfish Regions"}

\item{poly_name}{String Identifying the shapefile name that was used as mask.
Used to build file name.}
}
\value{
Time series dataframe for the selected region.
}
\description{
Tool for accessing oisst regional timeseries from Box.
Regionally masked timeseries have been pre-processed and stored on box. Quick access
is available here.

Simply specify your personal path to the oisst_mainstays folder,
the group of polygons your area is in, and the name of the region itself.
}
\examples{
#Not run:

# box_paths <- gmRi::research_access_paths(os.use = "unix",
#                                          user.name = "not applicable")
# agulhas <- oisst_access_timeseries(res_path = box_paths$oisst_mainstays,
#                                    region_family = "lme",
#                                    poly_name = "agulhas current")
}
