% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oisst_data_funs.R
\name{oisst_list_to_stack}
\alias{oisst_list_to_stack}
\title{Download OISST from Thredds, convert to Stack}
\usage{
oisst_list_to_stack(box, times.window)
}
\arguments{
\item{box}{bounding box object used to clip netcdf data}

\item{times.window}{two element vector of desired start and end dates.
Must be able to convert to date using as.Date}
}
\value{
rast.temp Raster stack of OI Sea Surface Temperatures
}
\description{
This function loops over daily OISST files online at
[ESRL](https://www.esrl.noaa.gov/psd/thredds/dodsC/Datasets/noaa.oisst.v2.highres/),
then downloads and returns a single raster stack.

Installs "ncdf4" and "raster" packages if not installed.
}
